pyproject.toml
[project]
name = "ghst"
version = "0.1.0"
description = "Shows your GitHub Starred project releases"
readme = "README.md"
authors = [{ name = "Fredrik Mellström", email = "fredrikm@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "pygithub>=2.5.0",
    "rich>=13.9.4",
    "typer>=0.15.1",
    "tzlocal>=5.2",
]

[dependency-groups]
dev = ["pylint>=3.3.2"]

[project.scripts]
ghst = "ghst:app.app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylint.main]
jobs = 0
py-version = "3.13"

[tool.pylint."messages control"]
disable = [
    "raw-checker-failed",
    "bad-inline-option",
    "locally-disabled",
    "file-ignored",
    "suppressed-message",
    "useless-suppression",
    "deprecated-pragma",
    "use-symbolic-message-instead",
    "use-implicit-booleaness-not-comparison-to-string",
    "use-implicit-booleaness-not-comparison-to-zero",
    "logging-fstring-interpolation",
]
# enable =

[tool.pylint.reports]
# Set the output format. Available formats are: text, parseable, colorized, json2
# (improved json format), json (old json format) and msvs (visual studio).
output-format = "colorized"
